Software_1.0_DataPipelines/US Market Research/requirements.txt
fredapi
matplotlib
pandas
python-dotenv
